ch2/ThoughtsBackend/requirements.txt
# Application requirements
flask==1.0.2
flask-restplus==0.12.1
flask-sqlalchemy==2.3.1
pytest-flask==0.14.0
freezegun==0.1.11
delorean==1.0.0
faker==1.0.7
